- customize google map as a widget.
