+++ lib/features/new_address/presentation/google_map_screen.dart
@@ -1,23 +1,16 @@
-import 'dart:async';
-import 'dart:collection';
-import 'dart:developer';
-import 'dart:math';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:test1/core/AppRouter.dart';
 import 'package:test1/core/constants/constants.dart';
 import 'package:test1/core/constants/vertical_and_horizontal_space.dart';
 import 'package:test1/core/helper/extensions.dart';
 import 'package:test1/core/theming/styles.dart';
-import 'package:test1/cubit/generic_cubit/generic_cubit.dart';
-import 'package:test1/cubit/generic_cubit/generic_state.dart';
-import 'package:test1/features/new_address/presentation/google_maps_view_model.dart';
-import 'package:test1/features/new_address/presentation/new_address_view_model.dart';
 import 'package:test1/features/shared/next_button.dart';
+
 import '../../../components/widgets/loader.dart';
 import '../../../core/widgets/custom_app_bar.dart';
 import '../../../core/widgets/custom_button.dart';
+import 'components/custom_google_map_widget.dart';
 
 class PolygonMapScreen extends StatefulWidget {
   final List<LatLng> points;
@@ -32,33 +25,6 @@ class PolygonMapScreen extends StatefulWidget {
 }
 
 class _PolygonMapScreenState extends State<PolygonMapScreen> {
-  GoogleMapsViewModel googleMapsViewModel = GoogleMapsViewModel();
-  final Completer<GoogleMapController> _controller = Completer();
-  late CameraPosition kGoogle;
-  final Set<Polygon> _polygon = HashSet<Polygon>();
-
-  List<LatLng> points = [];
-
-  @override
-  void initState() {
-    points = widget.points;
-    super.initState();
-    kGoogle = CameraPosition(
-      target: widget.points.first,
-      zoom: 15,
-    );
-    _polygon.add(
-      Polygon(
-        polygonId: const PolygonId('1'),
-        points: points,
-        fillColor: Colors.green.withOpacity(0.3),
-        strokeColor: Colors.green,
-        geodesic: true,
-        strokeWidth: 4,
-      ),
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
     return Directionality(
@@ -75,38 +41,8 @@ class _PolygonMapScreenState extends State<PolygonMapScreen> {
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.center,
                 children: [
-                  BlocBuilder<GenericCubit, GenericState>(
-                    bloc: googleMapsViewModel.googleMapsCubit,
-                    builder: (context, state) {
-                      return Expanded(
-                        child: GoogleMap(
-                          initialCameraPosition: kGoogle,
-                          mapType: MapType.normal,
-                          myLocationEnabled: true,
-                          myLocationButtonEnabled: true,
-                          compassEnabled: true,
-                          polygons: _polygon,
-                          markers: googleMapsViewModel.markers,
-                          // Add markers to the map
-                          onMapCreated: (GoogleMapController controller) {
-                            _controller.complete(controller);
-                          },
-                          onTap: (LatLng tappedPoint) {
-                            if (googleMapsViewModel.isPointInPolygon(
-                                tappedPoint, points)) {
-                              googleMapsViewModel.addMarker(
-                                  tappedPoint); // Add marker if within polygon
-                            } else {
-                              ScaffoldMessenger.of(context).showSnackBar(
-                                const SnackBar(
-                                    content: Text(
-                                        'يجب اختيار نقطة داخل المنطقة المحددة')),
-                              );
-                            }
-                          },
-                        ),
-                      );
-                    },
+                  CustomGoogleMap(
+                    points: widget.points,
                   ),
                   const VerticalSpacer(21),
                   Row(
